fix: double delete confirm
